EMBRYO-32: Ship2 renamed to Ship. Simple input valideringer tilføjet til ny ship information form.

+++ src/main/java/dk/dma/embryo/rest/json/Ship.java
@@ -41,7 +41,7 @@ public class Ship {
     
     private Integer length;
     
-    private Integer maxSpeed;
+    private Short maxSpeed;
 
     private Integer tonnage;
 
@@ -139,11 +139,11 @@ public class Ship {
         this.length = length;
     }
 
-    public Integer getMaxSpeed() {
+    public Short getMaxSpeed() {
         return maxSpeed;
     }
 
-    public void setMaxSpeed(Integer maxSpeed) {
+    public void setMaxSpeed(Short maxSpeed) {
         this.maxSpeed = maxSpeed;
     }
